Task 1 - Super Quick Update
Just changed the == to is, since we can do so because None is stored at one place in memory, so instead performing an equality check between whatever values I had and None, it's simpler to check if it references the same place in memory, as it's quicker. Nothing else was changed besides this.